Keep track of when a fan last read chat messages.

+++ src/main/java/se/artcomputer/f1/bingo/controller/ChatMessageRequest.java
@@ -1,4 +1,4 @@
 package se.artcomputer.f1.bingo.controller;
 
-public record ChatMessageRequest(String message, int fan) {
+public record ChatMessageRequest(String message, long fan) {
 }
